Now grabs tracks from Soundcloud and adds to redux store.
