+++ src/containers/SCButton.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { getTracksFromSC } from '../actions';
+import { getTracks } from '../actions';
 
 import scButton from '../assets/btn-connect-sc-l.png';
 
 const SCButton = props => {
   return (
     <button
-      onClick={props.getTracksFromSC}
+      onClick={props.getTracks}
       style={{ border: 'none', backgroundColor: 'white' }}
     >
       <img src={scButton} className="sc-button" alt="Connect with SoundCloud" />
@@ -20,7 +20,7 @@ const mapStateToProps = state => {
 };
 
 const mapDispatchToProps = dispatch => {
-  return { getTracksFromSC: () => dispatch(getTracksFromSC()) };
+  return { getTracks: () => dispatch(getTracks()) };
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(SCButton);
