+++ src/containers/TrackList.js
@@ -1,14 +1,17 @@
 import React from 'react';
+import { connect } from 'react-redux';
 import TrackListItem from '../components/TrackListItem';
 
-import { exampleTracks } from '../exampleTracks';
-
 const TrackList = props => {
-  let items = exampleTracks.map((track, i) => {
+  let items = props.trackList.map((track, i) => {
     return <TrackListItem track={track} key={i} />;
   });
 
   return <div>{items}</div>;
 };
 
-export default TrackList;
+const mapStateToProps = state => {
+  return { trackList: state.tracks.trackList };
+};
+
+export default connect(mapStateToProps)(TrackList);
